+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { DragZones } from "./components";
-import Button from "./components/Button";
+import { DragZones, Button } from "./components";
 
 const App: React.FC = () => {
   return (
