+++ src/components/DragZones/DragZones.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import DropZone from "./DropZone";
-import Button from "./Button";
-import Table from "./Table";
+import {DropZone ,Button , Table } from "../../components";
 import { useSelector } from "react-redux";
+import { useDragDrop } from "../../hooks/useDragDrop";
+
 const DragZones: React.FC = () => {
   const { title, description } = useSelector((state: any) => state.inputs);
   const { firstDrop, secondDrop, firstZoneComponent, secondZoneComponent } =
