+++ src/components/index.ts
@@ -3,4 +3,4 @@ export { default as TextInput } from "./TextInput/TextInput";
 export { default as Button } from "./Button/Button";
 export { default as TextArea } from "./TextArea/TextArea";
 export { default as Table } from "./Table/Table";
-export { default as DragZone } from "./DragZone/DragZone";
+export { default as DropZone } from "./DropZone/DropZone";
